[MNG-8151] Merge DependencyCollector into DependencyResolver
