+++ api/maven-api-core/src/main/java/org/apache/maven/api/services/DependencyResolver.java
@@ -20,6 +20,7 @@ package org.apache.maven.api.services;
 
 import java.util.List;
 
+import org.apache.maven.api.Artifact;
 import org.apache.maven.api.DependencyCoordinate;
 import org.apache.maven.api.Node;
 import org.apache.maven.api.PathScope;
@@ -35,55 +36,165 @@ import org.apache.maven.api.annotations.Nonnull;
 @Experimental
 public interface DependencyResolver extends Service {
 
+    /**
+     * Collects the transitive dependencies of some artifacts and builds a dependency graph. Note that this operation is
+     * only concerned about determining the coordinates of the transitive dependencies and does not actually resolve the
+     * artifact files.
+     *
+     * @param session the {@link Session}, must not be {@code null}
+     * @param root the Maven Dependency, must not be {@code null}
+     * @return the collection result, never {@code null}
+     * @throws DependencyResolverException if the dependency tree could not be built
+     * @throws IllegalArgumentException if an argument is null or invalid
+     * @see #collect(DependencyResolverRequest)
+     */
+    @Nonnull
+    default DependencyResolverResult collect(@Nonnull Session session, @Nonnull DependencyCoordinate root) {
+        return collect(DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.COLLECT, root));
+    }
+
+    /**
+     * Collects the transitive dependencies of some artifacts and builds a dependency graph. Note that this operation is
+     * only concerned about determining the coordinates of the transitive dependencies and does not actually resolve the
+     * artifact files.
+     *
+     * @param session the {@link Session}, must not be {@code null}
+     * @param project the {@link Project}, must not be {@code null}
+     * @return the collection result, never {@code null}
+     * @throws DependencyResolverException if the dependency tree could not be built
+     * @throws IllegalArgumentException if an argument is null or invalid
+     * @see #collect(DependencyResolverRequest)
+     */
+    @Nonnull
+    default DependencyResolverResult collect(@Nonnull Session session, @Nonnull Project project) {
+        return collect(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.COLLECT, project));
+    }
+
+    /**
+     * Collects the transitive dependencies of some artifacts and builds a dependency graph. Note that this operation is
+     * only concerned about determining the coordinates of the transitive dependencies and does not actually resolve the
+     * artifact files.
+     *
+     * @param session the {@link Session}, must not be {@code null}
+     * @param artifact the {@link Artifact}, must not be {@code null}
+     * @return the collection result, never {@code null}
+     * @throws DependencyResolverException if the dependency tree could not be built
+     * @throws IllegalArgumentException if an argument is null or invalid
+     * @see #collect(DependencyResolverRequest)
+     */
+    @Nonnull
+    default DependencyResolverResult collect(@Nonnull Session session, @Nonnull Artifact artifact) {
+        return collect(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.COLLECT, artifact));
+    }
+
+    /**
+     * Collects the transitive dependencies and builds a dependency graph.
+     * Note that this operation is only concerned about determining the coordinates of the
+     * transitive dependencies and does not actually resolve the artifact files.
+     *
+     * @param request the dependency collection request, must not be {@code null}
+     * @return the collection result, never {@code null}
+     * @throws DependencyResolverException if the dependency tree could not be built
+     * @throws IllegalArgumentException if an argument is null or invalid
+     *
+     * @see DependencyResolver#collect(Session, Project)
+     * @see DependencyResolver#collect(Session, DependencyCoordinate)
+     * @see DependencyResolver#collect(Session, Artifact)
+     */
+    @Nonnull
+    default DependencyResolverResult collect(@Nonnull DependencyResolverRequest request) {
+        if (request.getRequestType() != DependencyResolverRequest.RequestType.COLLECT) {
+            throw new IllegalArgumentException("requestType should be COLLECT when calling collect()");
+        }
+        return resolve(request);
+    }
+
+    /**
+     * Flattens a list of nodes.
+     *
+     * @param session
+     * @param node
+     * @param scope
+     * @return
+     * @throws DependencyResolverException
+     */
     List<Node> flatten(Session session, Node node, PathScope scope) throws DependencyResolverException;
+
+    @Nonnull
+    default DependencyResolverResult flatten(@Nonnull Session session, @Nonnull Project project) {
+        return flatten(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.FLATTEN, project));
+    }
+
+    @Nonnull
+    default DependencyResolverResult flatten(
+            @Nonnull Session session, @Nonnull Project project, @Nonnull PathScope scope) {
+        return flatten(DependencyResolverRequest.build(
+                session, DependencyResolverRequest.RequestType.FLATTEN, project, scope));
+    }
+
+    @Nonnull
+    default DependencyResolverResult flatten(@Nonnull DependencyResolverRequest request) {
+        if (request.getRequestType() != DependencyResolverRequest.RequestType.FLATTEN) {
+            throw new IllegalArgumentException("requestType should be FLATTEN when calling flatten()");
+        }
+        return resolve(request);
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(@Nonnull Session session, @Nonnull Project project) {
+        return resolve(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.RESOLVE, project));
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(
+            @Nonnull Session session, @Nonnull Project project, @Nonnull PathScope scope) {
+        return resolve(DependencyResolverRequest.build(
+                session, DependencyResolverRequest.RequestType.RESOLVE, project, scope));
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(@Nonnull Session session, @Nonnull DependencyCoordinate dependency) {
+        return resolve(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.RESOLVE, dependency));
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(
+            @Nonnull Session session, @Nonnull DependencyCoordinate dependency, @Nonnull PathScope scope) {
+        return resolve(DependencyResolverRequest.build(
+                session, DependencyResolverRequest.RequestType.RESOLVE, dependency, scope));
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(
+            @Nonnull Session session, @Nonnull List<DependencyCoordinate> dependencies) {
+        return resolve(
+                DependencyResolverRequest.build(session, DependencyResolverRequest.RequestType.RESOLVE, dependencies));
+    }
+
+    @Nonnull
+    default DependencyResolverResult resolve(
+            @Nonnull Session session, @Nonnull List<DependencyCoordinate> dependencies, @Nonnull PathScope scope) {
+        return resolve(DependencyResolverRequest.build(
+                session, DependencyResolverRequest.RequestType.RESOLVE, dependencies, scope));
+    }
 
     /**
      * This method collects, flattens and resolves the dependencies.
      *
      * @param request the request to resolve
      * @return the result of the resolution
-     * @throws DependencyCollectorException
      * @throws DependencyResolverException
      * @throws ArtifactResolverException
      *
-     * @see DependencyCollector#collect(DependencyCollectorRequest)
+     * @see DependencyResolver#collect(DependencyResolverRequest)
      * @see #flatten(Session, Node, PathScope)
      * @see ArtifactResolver#resolve(ArtifactResolverRequest)
      */
     DependencyResolverResult resolve(DependencyResolverRequest request)
-            throws DependencyCollectorException, DependencyResolverException, ArtifactResolverException;
-
-    @Nonnull
-    default DependencyResolverResult resolve(@Nonnull Session session, @Nonnull Project project) {
-        return resolve(DependencyResolverRequest.build(session, project));
-    }
-
-    @Nonnull
-    default DependencyResolverResult resolve(
-            @Nonnull Session session, @Nonnull Project project, @Nonnull PathScope scope) {
-        return resolve(DependencyResolverRequest.build(session, project, scope));
-    }
-
-    @Nonnull
-    default DependencyResolverResult resolve(@Nonnull Session session, @Nonnull DependencyCoordinate dependency) {
-        return resolve(DependencyResolverRequest.build(session, dependency));
-    }
-
-    @Nonnull
-    default DependencyResolverResult resolve(
-            @Nonnull Session session, @Nonnull DependencyCoordinate dependency, @Nonnull PathScope scope) {
-        return resolve(DependencyResolverRequest.build(session, dependency, scope));
-    }
-
-    @Nonnull
-    default DependencyResolverResult resolve(
-            @Nonnull Session session, @Nonnull List<DependencyCoordinate> dependencies) {
-        return resolve(DependencyResolverRequest.build(session, dependencies));
-    }
-
-    @Nonnull
-    default DependencyResolverResult resolve(
-            @Nonnull Session session, @Nonnull List<DependencyCoordinate> dependencies, @Nonnull PathScope scope) {
-        return resolve(DependencyResolverRequest.build(session, dependencies, scope));
-    }
+            throws DependencyResolverException, ArtifactResolverException;
 }
