+++ api/maven-api-core/src/main/java/org/apache/maven/api/services/DependencyResolverRequest.java
@@ -18,8 +18,11 @@
  */
 package org.apache.maven.api.services;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Predicate;
 
 import org.apache.maven.api.Artifact;
@@ -30,14 +33,54 @@ import org.apache.maven.api.PathType;
 import org.apache.maven.api.Project;
 import org.apache.maven.api.Session;
 import org.apache.maven.api.annotations.Experimental;
+import org.apache.maven.api.annotations.Immutable;
 import org.apache.maven.api.annotations.Nonnull;
 import org.apache.maven.api.annotations.NotThreadSafe;
 import org.apache.maven.api.annotations.Nullable;
 
+/**
+ * A request to collect the transitive dependencies and to build a dependency graph from them. There are three ways to
+ * create a dependency graph. First, only the root dependency can be given. Second, a root dependency and direct
+ * dependencies can be specified in which case the specified direct dependencies are merged with the direct dependencies
+ * retrieved from the artifact descriptor of the root dependency. And last, only direct dependencies can be specified in
+ * which case the root node of the resulting graph has no associated dependency.
+ *
+ * @since 4.0.0
+ */
 @Experimental
-public interface DependencyResolverRequest extends DependencyCollectorRequest {
+@Immutable
+public interface DependencyResolverRequest {
+
+    enum RequestType {
+        COLLECT,
+        FLATTEN,
+        RESOLVE
+    }
 
     @Nonnull
+    Session getSession();
+
+    @Nonnull
+    RequestType getRequestType();
+
+    @Nonnull
+    Optional<Project> getProject();
+
+    @Nonnull
+    Optional<Artifact> getRootArtifact();
+
+    @Nonnull
+    Optional<DependencyCoordinate> getRoot();
+
+    @Nonnull
+    Collection<DependencyCoordinate> getDependencies();
+
+    @Nonnull
+    Collection<DependencyCoordinate> getManagedDependencies();
+
+    boolean getVerbose();
+
+    @Nullable
     PathScope getPathScope();
 
     /**
@@ -48,6 +91,7 @@ public interface DependencyResolverRequest extends DependencyCollectorRequest {
      *
      * @return a filter for the types of path (class-path, module-path, …) accepted by the tool
      */
+    @Nullable
     Predicate<PathType> getPathTypeFilter();
 
     @Nonnull
@@ -56,119 +100,202 @@ public interface DependencyResolverRequest extends DependencyCollectorRequest {
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, Project project) {
-        return build(session, project, PathScope.MAIN_RUNTIME);
+    static DependencyResolverRequest build(Session session, RequestType requestType, Artifact rootArtifact) {
+        return new DependencyResolverRequestBuilder()
+                .session(session)
+                .requestType(requestType)
+                .rootArtifact(rootArtifact)
+                .pathScope(PathScope.MAIN_RUNTIME)
+                .build();
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, Project project, PathScope scope) {
+    static DependencyResolverRequest build(Session session, RequestType requestType, Project project) {
+        return build(session, requestType, project, PathScope.MAIN_RUNTIME);
+    }
+
+    @Nonnull
+    static DependencyResolverRequest build(Session session, RequestType requestType, Project project, PathScope scope) {
         return new DependencyResolverRequestBuilder()
                 .session(session)
+                .requestType(requestType)
                 .project(project)
                 .pathScope(scope)
                 .build();
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, DependencyCoordinate dependency) {
-        return build(session, dependency, PathScope.MAIN_RUNTIME);
+    static DependencyResolverRequest build(Session session, RequestType requestType, DependencyCoordinate dependency) {
+        return build(session, requestType, dependency, PathScope.MAIN_RUNTIME);
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, DependencyCoordinate dependency, PathScope scope) {
+    static DependencyResolverRequest build(
+            Session session, RequestType requestType, DependencyCoordinate dependency, PathScope scope) {
         return new DependencyResolverRequestBuilder()
                 .session(session)
+                .requestType(requestType)
                 .dependency(dependency)
                 .pathScope(scope)
                 .build();
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, List<DependencyCoordinate> dependencies) {
-        return build(session, dependencies, PathScope.MAIN_RUNTIME);
+    static DependencyResolverRequest build(
+            Session session, RequestType requestType, List<DependencyCoordinate> dependencies) {
+        return build(session, requestType, dependencies, PathScope.MAIN_RUNTIME);
     }
 
     @Nonnull
-    static DependencyResolverRequest build(Session session, List<DependencyCoordinate> dependencies, PathScope scope) {
+    static DependencyResolverRequest build(
+            Session session, RequestType requestType, List<DependencyCoordinate> dependencies, PathScope scope) {
         return new DependencyResolverRequestBuilder()
                 .session(session)
+                .requestType(requestType)
                 .dependencies(dependencies)
                 .pathScope(scope)
                 .build();
     }
 
     @NotThreadSafe
-    class DependencyResolverRequestBuilder extends DependencyCollectorRequestBuilder {
-        PathScope pathScope;
+    class DependencyResolverRequestBuilder {
 
+        Session session;
+        RequestType requestType;
+        Project project;
+        Artifact rootArtifact;
+        DependencyCoordinate root;
+        List<DependencyCoordinate> dependencies = Collections.emptyList();
+        List<DependencyCoordinate> managedDependencies = Collections.emptyList();
+        boolean verbose;
+        PathScope pathScope;
         Predicate<PathType> pathTypeFilter;
 
+        DependencyResolverRequestBuilder() {}
+
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder session(@Nonnull Session session) {
-            super.session(session);
+            this.session = session;
             return this;
         }
 
         @Nonnull
-        @Override
+        public DependencyResolverRequestBuilder requestType(@Nonnull RequestType requestType) {
+            this.requestType = requestType;
+            return this;
+        }
+
+        @Nonnull
         public DependencyResolverRequestBuilder project(@Nullable Project project) {
-            super.project(project);
+            this.project = project;
             return this;
         }
 
+        /**
+         * Sets the root artifact for the dependency graph.
+         * This must not be confused with {@link #root(DependencyCoordinate)}: The root <em>dependency</em>, like any
+         * other specified dependency, will be subject to dependency collection/resolution, i.e. should have an artifact
+         * descriptor and a corresponding artifact file. The root <em>artifact</em> on the other hand is only used
+         * as a label for the root node of the graph in case no root dependency was specified. As such, the configured
+         * root artifact is ignored if {@link #root(DependencyCoordinate)} has been set.
+         *
+         * @param rootArtifact the root artifact for the dependency graph, may be {@code null}
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder rootArtifact(@Nullable Artifact rootArtifact) {
-            super.rootArtifact(rootArtifact);
+            this.rootArtifact = rootArtifact;
             return this;
         }
 
+        /**
+         * @param root The root dependency
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
-        public DependencyResolverRequestBuilder root(@Nullable DependencyCoordinate root) {
-            super.root(root);
+        public DependencyResolverRequestBuilder root(@Nonnull DependencyCoordinate root) {
+            this.root = root;
             return this;
         }
 
+        /**
+         * Sets the direct dependencies. If both a root dependency and direct dependencies are given in the request, the
+         * direct dependencies from the request will be merged with the direct dependencies from the root dependency's
+         * artifact descriptor, giving higher priority to the dependencies from the request.
+         *
+         * @param dependencies the direct dependencies, may be {@code null}
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder dependencies(@Nullable List<DependencyCoordinate> dependencies) {
-            super.dependencies(dependencies);
+            this.dependencies = (dependencies != null) ? dependencies : Collections.emptyList();
             return this;
         }
 
+        /**
+         * Adds the specified direct dependency.
+         *
+         * @param dependency the dependency to add, may be {@code null}
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder dependency(@Nullable DependencyCoordinate dependency) {
-            super.dependency(dependency);
+            if (dependency != null) {
+                if (this.dependencies.isEmpty()) {
+                    this.dependencies = new ArrayList<>();
+                }
+                this.dependencies.add(dependency);
+            }
             return this;
         }
 
+        /**
+         * Sets the dependency management to apply to transitive dependencies. To clarify, this management does not
+         * apply to
+         * the direct dependencies of the root node.
+         *
+         * @param managedDependencies the dependency management, may be {@code null}
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder managedDependencies(
                 @Nullable List<DependencyCoordinate> managedDependencies) {
-            super.managedDependencies(managedDependencies);
+            this.managedDependencies = (managedDependencies != null) ? managedDependencies : Collections.emptyList();
             return this;
         }
 
+        /**
+         * Adds the specified managed dependency.
+         *
+         * @param managedDependency The managed dependency to add, may be {@code null} in which case the call
+         *                          will have no effect.
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder managedDependency(@Nullable DependencyCoordinate managedDependency) {
-            super.managedDependency(managedDependency);
+            if (managedDependency != null) {
+                if (this.managedDependencies.isEmpty()) {
+                    this.managedDependencies = new ArrayList<>();
+                }
+                this.managedDependencies.add(managedDependency);
+            }
             return this;
         }
 
+        /**
+         * Specifies that the collection should be verbose.
+         *
+         * @param verbose whether the collection should be verbose or not
+         * @return this request for chaining, never {@code null}
+         */
         @Nonnull
-        @Override
         public DependencyResolverRequestBuilder verbose(boolean verbose) {
-            super.verbose(verbose);
+            this.verbose = verbose;
             return this;
         }
 
         @Nonnull
-        public DependencyResolverRequestBuilder pathScope(@Nonnull PathScope pathScope) {
+        public DependencyResolverRequestBuilder pathScope(@Nullable PathScope pathScope) {
             this.pathScope = pathScope;
             return this;
         }
@@ -200,10 +327,11 @@ public interface DependencyResolverRequest extends DependencyCollectorRequest {
             return pathTypeFilter(desiredTypes::contains);
         }
 
-        @Override
+        @Nonnull
         public DependencyResolverRequest build() {
             return new DefaultDependencyResolverRequest(
                     session,
+                    requestType,
                     project,
                     rootArtifact,
                     root,
@@ -214,31 +342,92 @@ public interface DependencyResolverRequest extends DependencyCollectorRequest {
                     pathTypeFilter);
         }
 
-        static class DefaultDependencyResolverRequest extends DefaultDependencyCollectorRequest
-                implements DependencyResolverRequest {
+        static class DefaultDependencyResolverRequest extends BaseRequest implements DependencyResolverRequest {
+            private final RequestType requestType;
+            private final Project project;
+            private final Artifact rootArtifact;
+            private final DependencyCoordinate root;
+            private final Collection<DependencyCoordinate> dependencies;
+            private final Collection<DependencyCoordinate> managedDependencies;
+            private final boolean verbose;
             private final PathScope pathScope;
-
             private final Predicate<PathType> pathTypeFilter;
 
+            /**
+             * Creates a request with the specified properties.
+             *
+             * @param session      {@link Session}
+             * @param rootArtifact The root dependency whose transitive dependencies should be collected, may be {@code
+             *                     null}.
+             */
             DefaultDependencyResolverRequest(
-                    Session session,
-                    Project project,
-                    Artifact rootArtifact,
-                    DependencyCoordinate root,
-                    Collection<DependencyCoordinate> dependencies,
-                    Collection<DependencyCoordinate> managedDependencies,
+                    @Nonnull Session session,
+                    @Nonnull RequestType requestType,
+                    @Nullable Project project,
+                    @Nullable Artifact rootArtifact,
+                    @Nullable DependencyCoordinate root,
+                    @Nonnull Collection<DependencyCoordinate> dependencies,
+                    @Nonnull Collection<DependencyCoordinate> managedDependencies,
                     boolean verbose,
-                    PathScope pathScope,
-                    Predicate<PathType> pathTypeFilter) {
-                super(session, project, rootArtifact, root, dependencies, managedDependencies, verbose);
-                this.pathScope = nonNull(pathScope, "pathScope cannot be null");
+                    @Nullable PathScope pathScope,
+                    @Nullable Predicate<PathType> pathTypeFilter) {
+                super(session);
+                this.requestType = nonNull(requestType, "requestType cannot be null");
+                this.project = project;
+                this.rootArtifact = rootArtifact;
+                this.root = root;
+                this.dependencies = unmodifiable(nonNull(dependencies, "dependencies cannot be null"));
+                this.managedDependencies =
+                        unmodifiable(nonNull(managedDependencies, "managedDependencies cannot be null"));
+                this.verbose = verbose;
+                this.pathScope = pathScope;
                 this.pathTypeFilter = (pathTypeFilter != null) ? pathTypeFilter : (t) -> true;
-                if (verbose) {
-                    throw new IllegalArgumentException("verbose cannot be true for resolving dependencies");
+                if (verbose && requestType != RequestType.COLLECT) {
+                    throw new IllegalArgumentException("verbose cannot only be true when collecting dependencies");
                 }
             }
 
             @Nonnull
+            @Override
+            public RequestType getRequestType() {
+                return requestType;
+            }
+
+            @Nonnull
+            @Override
+            public Optional<Project> getProject() {
+                return Optional.ofNullable(project);
+            }
+
+            @Nonnull
+            @Override
+            public Optional<Artifact> getRootArtifact() {
+                return Optional.ofNullable(rootArtifact);
+            }
+
+            @Nonnull
+            @Override
+            public Optional<DependencyCoordinate> getRoot() {
+                return Optional.ofNullable(root);
+            }
+
+            @Nonnull
+            @Override
+            public Collection<DependencyCoordinate> getDependencies() {
+                return dependencies;
+            }
+
+            @Nonnull
+            @Override
+            public Collection<DependencyCoordinate> getManagedDependencies() {
+                return managedDependencies;
+            }
+
+            @Override
+            public boolean getVerbose() {
+                return verbose;
+            }
+
             @Override
             public PathScope getPathScope() {
                 return pathScope;
@@ -247,6 +436,12 @@ public interface DependencyResolverRequest extends DependencyCollectorRequest {
             @Override
             public Predicate<PathType> getPathTypeFilter() {
                 return pathTypeFilter;
+            }
+
+            @Nonnull
+            @Override
+            public String toString() {
+                return getRoot() + " -> " + getDependencies();
             }
         }
     }
